refactor: Modularize application logic
Separate the application logic into distinct modules for better organization and maintainability.

- Create 'win' module for window management.
- Create 'gate' module for application-specific logic.
- Update 'main' to orchestrate the modules.
- Add all new source files to the Makefile.
