+++ gate.h
@@ -1,0 +1,4 @@
+#ifndef GATE_H_
+#define GATE_H_
+int gate_initialize(t_data *data);
+#endif//GATE_H_
