+++ win.h
@@ -1,0 +1,8 @@
+#ifndef WIN_H_
+#define WIN_H_
+#include "main.h"
+
+int win_initialize(t_data *data);
+void win_loop(t_data *data);
+
+#endif//WIN_H_
